tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/keyframes_not_cascading.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - cascading within a @keyframes rule</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#keyframes">
   <meta name="flags" content="dom">
   <meta name="assert" content="There is no cascading within a @keyframes rule
   if multiple keyframes specify the same keyframe selector values">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            left: 0px;
         }
         100% {
            left: 100px;
         }
         to {
            top: 0px;
         }
         0% {
            top: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: slide 1s forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var startOK = false;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function() {
            var topInPx = parseFloat(getComputedStyle(div, null).top);
            startOK = (topInPx > 50 && topInPx <= 100
                       && getComputedStyle(div, null).left == 'auto');
         });
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            var s = getComputedStyle(div, null);
            var passed = (startOK && s.top == '0px' && s.left == 'auto');
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should slide upwards without moving sideways.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
